requirements.txt
setuptools~=59.1.1

requests==2.26.0
beautifulsoup4==4.10.0
python-magic==0.4.25
